refactor(furniture): split serializers into separate modules
