+++ furniture/serializers/__init__.py
@@ -1,0 +1,6 @@
+from .base import AbstractCoordinates, CoordinateSerializer
+from .placements import (DoorPlacementSerializer, FurniturePlacementSerializer,
+                         PowerSocketPlacementSerializer,
+                         WindowPlacementSerializer)
+from .room_items import FurnitureSerializer, RoomTypeSerializer
+from .room_layout import RoomLayoutCopySerializer, RoomLayoutSerializer
